Double precision (#895)
* added configuration summary

* use custom math type

* use std function if possible

* build success

* fix js binding

* format

* use better primitives

* use old precision for polygon tests

* fix polygon CCW check

* temp-fix tests

* add simpler failing test case

* refactored CutKeyhole

* fixed tests

* more keyholing fixes

* truncate long debug output

* shrink SDF blobs test

* fixed WASM and disabled zebra

* fix hull

* cleanup a bit

* fix wasm

* CI cleanuo

* reenable zebra

* upped precision and fixed tests

* fix format

* disable zebra test on windows

---------

+++ bindings/c/meshio.cpp
@@ -45,11 +45,11 @@ ManifoldMaterial *manifold_material(void *mem) {
   return to_c(new (mem) manifold::Material());
 }
 
-void manifold_material_set_roughness(ManifoldMaterial *mat, float roughness) {
+void manifold_material_set_roughness(ManifoldMaterial *mat, double roughness) {
   from_c(mat)->roughness = roughness;
 }
 
-void manifold_material_set_metalness(ManifoldMaterial *mat, float metalness) {
+void manifold_material_set_metalness(ManifoldMaterial *mat, double metalness) {
   from_c(mat)->metalness = metalness;
 }
